добавлены интерфейсы сервисов
